rename

+++ neo/SmartContract/Native/Tokens/GasToken.cs
@@ -15,7 +15,7 @@ namespace Neo.SmartContract.Native.Tokens
         public override string ServiceName => "Neo.Native.Tokens.GAS";
         public override string Name => "GAS";
         public override string Symbol => "gas";
-        public override int Decimals => 8;
+        public override byte Decimals => 8;
 
         private const byte Prefix_SystemFeeAmount = 15;
